+++ miniapp/src/main/java/com/rakuten/tech/mobile/miniapp/js/Type.kt
@@ -12,6 +12,7 @@ internal enum class ActionType(val action: String) {
     GET_ACCESS_TOKEN("getAccessToken"),
     SET_SCREEN_ORIENTATION("setScreenOrientation"),
     GET_CONTACTS("getContacts"),
+    SEND_MESSAGE_TO_CONTACT("sendMessageToContact"),
 }
 
 internal enum class DialogType {